name comparaison in another file
